+++ src/components/domain/Post/Bookmark.js
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled'
+
+const Bookmark = styled.button`
+  position: absolute;
+  top: -24px;
+  right: 40px;
+`
+
+export default Bookmark
